# auth/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This auth API only functions with OAuth2 and it shares the same database with [...]",
        "title": "Authentication REST API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/authenticate/{token}": {
            "get": {
                "description": "This handles both user creation and user authentication using OAuth2, via the google provider,",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authenticate"
                ],
                "summary": "User authentication",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/authenticate.response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/app.BaseErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.BaseErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "app.BaseErrorResponse": {
            "type": "object",
            "required": [
                "error"
            ],
            "properties": {
                "error": {
                    "type": "string"
                },
                "ok": {
                    "type": "boolean"
                }
            }
        },
        "authenticate.response": {
            "type": "object",
            "required": [
                "user"
            ],
            "properties": {
                "ok": {
                    "type": "boolean"
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                }
            }
        },
        "model.User": {
            "type": "object",
            "required": [
                "email",
                "id",
                "name"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "profilePictureUrl": {
                    "type": "string"
                }
            }
        }
    }
}